First step to convert ImmutableList to LinkedList.
ImmutableList extends LinkedList and reimplemented all the methods such
that the order and usages of ImmutableList is kept the same.

+++ core/src/test/java/io/parsingdata/metal/CurrentOffsetTest.java
@@ -53,8 +53,8 @@ public class CurrentOffsetTest {
         final ImmutableList<Value> offset = CURRENT_OFFSET.eval(result.get(), enc());
 
         assertNotNull(offset);
-        assertEquals(1, offset.size);
-        assertEquals(size, offset.head.asNumeric().longValueExact());
+        assertEquals(1, (long) offset.size());
+        assertEquals(size, offset.head().asNumeric().longValueExact());
     }
 
     @Test
